yueBpw/vertex/secure-api/secure_api_requirements.txt
Flask==3.0.0
requests==2.31.0
google-auth==2.23.4
gunicorn==21.2.0 